fullstack-orthanc-dicom/master-data-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
gunicorn==22.0.0